requirements.txt
python-dotenv
selenium
python-telegram-bot

